requirements-assets.txt
# Solo per scripts/build-images.py (rigenera WebP/ICO da PNG sorgente)
pillow>=10.0.0
